* stm/impy: (ad5933.h) Fix wrong data type in range settings structure.

+++ trunk/stm/include/ad5933.h
@@ -41,8 +41,8 @@ typedef struct
 
 typedef struct
 {
-    uint8_t  PGA_Gain;          //!< PGA gain setting (one of the {@link AD5933_GAIN} values)
-    uint8_t  Voltage_Range;     //!< Voltage range setting (one of the {@link AD5933_VOLTAGE} values)
+    uint16_t  PGA_Gain;         //!< PGA gain setting (one of the {@link AD5933_GAIN} values)
+    uint16_t  Voltage_Range;    //!< Voltage range setting (one of the {@link AD5933_VOLTAGE} values)
 } AD5933_RangeSettings;
 
 typedef struct
